+++ package.json
@@ -21,6 +21,7 @@
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "gulp": "^4.0.2",
     "mongodb": "^5.6.0",
